requirements.txt
Flask
Werkzeug
python_jose
mysql-connector-python
requests requests_oauthlib